Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/zero-column-width-computed-style-expected.txt


FAIL column-width:0 is a valid CSS declaration assert_equals: expected "0px" but got "auto"
